Move traits to utility directory
